requirements.txt
Django==5.0.4
djangorestframework==3.15.1
python-dotenv==1.0.1
web3==6.18.0
eth-account==0.10.0
requests==2.31.0
Pillow==10.3.0
reportlab==4.2.0
qrcode==7.4.2
pinatapy-vourhey==0.1.9
PyPDF2==3.0.1
gunicorn==21.2.0
whitenoise==6.6.0

